-Fixed segfault on loading samples
This commit fixes an issue when loading new samples sometimes a segfault
in operator new() would occur. It is due to memory corruption from the
waveform caching code, and could be reproduced by loading a file, showing
the waveform in the UI by clicking the pad, and then loading another file.