Maschine OSC interface initial support
This commit adds the OscPkt header files, in order to have a
lightweight OSC server embedded in Fabla2. This allows UI commands
like changing tabs, browsing samples and changing parameters to be
performed via the Maschine hardware as well as on the software UI.

Currently F1 F2 and F3 buttons map to the three live view, and the
record button will record a new sample on the last played pad.