dsp: sequencing capabilities added
This commit takes the OpenAV sequencer code done for Buska, and integrates
it into Fabla2. Each "bank" (aka 1 drumkit) is given a Pattern class,
which owns 16 sequencers. These sequencers each drive one of the pads, and
to make a coherent whole drum-track.

Currently only one Pattern is supported per Bank, but this can (and should)
be extended to allow more flexibility for variations of beats.