ctlra: add data ring to pass arguments to function
This method uses a ringbuffer to pass the data required to
the function being moved between threads. The ringbuffer allows
variable-sized enqueues, and the message contained in the msg
ring has the size of data to be dequeued.

As such, the data pointer to the function is always the first
bytes available from the read size of the data ring - making it
obsolete to pass a pointer to the data itself.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>